merge

+++ script/Sh13_kool_proxy.sh
@@ -5,12 +5,32 @@ TAG="AD_BYBY"		  # iptables tag
 koolproxy_enable=`nvram get koolproxy_enable`
 [ -z $koolproxy_enable ] && koolproxy_enable=0 && nvram set koolproxy_enable=0
 if [ "$koolproxy_enable" != "0" ] ; then
-nvramshow=`nvram showall | grep '=' | grep ss | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
-nvramshow=`nvram showall | grep '=' | grep adbyby | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
-nvramshow=`nvram showall | grep '=' | grep adm | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
-nvramshow=`nvram showall | grep '=' | grep koolproxy | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
-
+#nvramshow=`nvram showall | grep '=' | grep ss | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
+#nvramshow=`nvram showall | grep '=' | grep adbyby | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
+#nvramshow=`nvram showall | grep '=' | grep adm | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
+#nvramshow=`nvram showall | grep '=' | grep koolproxy | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
+adbyby_mode_x=`nvram get adbyby_mode_x`
 [ -z $adbyby_mode_x ] && adbyby_mode_x=0 && nvram set adbyby_mode_x=0
+ss_link_1=`nvram get ss_link_1`
+koolproxy_auto=`nvram get koolproxy_auto`
+koolproxy_video=`nvram get koolproxy_video`
+koolproxy_update=`nvram get koolproxy_update`
+koolproxy_update_hour=`nvram get koolproxy_update_hour`
+koolproxy_update_min=`nvram get koolproxy_update_min`
+lan_ipaddr=`nvram get lan_ipaddr`
+koolproxy_https=`nvram get koolproxy_https`
+adm_hookport=`nvram get adm_hookport`
+koolproxy_adblock=`nvram get koolproxy_adblock`
+adbyby_CPUAverages=`nvram get adbyby_CPUAverages`
+ss_DNS_Redirect=`nvram get ss_DNS_Redirect`
+ss_DNS_Redirect_IP=`nvram get ss_DNS_Redirect_IP`
+ss_link_1=`nvram get ss_link_1`
+adbyby_enable=`nvram get adbyby_enable`
+adm_enable=`nvram get adm_enable`
+ss_enable=`nvram get ss_enable`
+ss_mode_x=`nvram get ss_mode_x`
+adbyby_adblocks=`nvram get adbyby_adblocks`
+koolproxy_uprules=`nvram get koolproxy_uprules`
 
 koolproxyfile="https://koolproxy.com/downloads/mipsel"
 koolproxyfilecdn="https://github.com/koolproxy/koolproxy-bin/raw/master/mipsel"
@@ -42,10 +62,11 @@ fi
 # echo "$mode_video"
 fi
 #检查 dnsmasq 目录参数
-confdir=`grep conf-dir /etc/storage/dnsmasq/dnsmasq.conf | sed 's/.*\=//g'`
-if [ -z "$confdir" ] ; then 
+#confdir=`grep "/tmp/ss/dnsmasq.d" /etc/storage/dnsmasq/dnsmasq.conf | sed 's/.*\=//g'`
+#if [ -z "$confdir" ] ; then 
 	confdir="/tmp/ss/dnsmasq.d"
-fi
+#fi
+confdir_x="$(echo -e $confdir | sed -e "s/\//"'\\'"\//g")"
 [ ! -d "$confdir" ] && mkdir -p $confdir
 gfwlist="/r.gfwlist.conf"
 gfw_black_list="gfwlist"
@@ -60,15 +81,28 @@ koolproxy_mount () {
 
 ss_opt_x=`nvram get ss_opt_x`
 upanPath=""
-[ "$ss_opt_x" = "3" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
-[ "$ss_opt_x" = "4" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
-[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
-[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
+[ "$ss_opt_x" = "3" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+[ "$ss_opt_x" = "4" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+if [ "$ss_opt_x" = "5" ] ; then
+	# 指定目录
+	opt_cifs_dir=`nvram get opt_cifs_dir`
+	if [ -d $opt_cifs_dir ] ; then
+		upanPath="$opt_cifs_dir"
+	else
+		logger -t "【opt】" "错误！未找到指定目录 $opt_cifs_dir"
+		upanPath=""
+		[ -z "$upanPath" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+		[ -z "$upanPath" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | sort -u | awk 'NR==1' `"
+	fi
+fi
 echo "$upanPath"
 if [ ! -z "$upanPath" ] ; then 
 	logger -t "【koolproxy】" "已挂载储存设备, 主程序放外置设备存储"
 	initopt
 	mkdir -p $upanPath/ad/7620koolproxy
+	rm -f /tmp/7620koolproxy
 	ln -sf "$upanPath/ad/7620koolproxy" /tmp/7620koolproxy
 	if [ -s /etc_ro/7620koolproxy_*.tgz ] && [ ! -s "$upanPath/ad/7620koolproxy/koolproxy" ] ; then
 		logger -t "【koolproxy】" "使用内置主程序"
@@ -93,7 +127,8 @@ else
 	fi
 fi
 export PATH='/tmp/7620koolproxy:/etc/storage/bin:/tmp/script:/etc/storage/script:/opt/usr/sbin:/opt/usr/bin:/opt/sbin:/opt/bin:/usr/local/sbin:/usr/sbin:/usr/bin:/sbin:/bin'
-hash koolproxy 2>/dev/null || rm -rf /tmp/7620koolproxy/*
+chmod 777 /tmp/7620koolproxy/koolproxy
+[[ "$(/tmp/7620koolproxy/koolproxy -h | wc -l)" -lt 2 ]] && rm -rf /tmp/7620koolproxy/*
 }
 
 koolproxy_restart () {
@@ -151,7 +186,7 @@ koolproxy_check () {
 koolproxy_get_status
 if [ "$koolproxy_enable" != "1" ] && [ "$needed_restart" = "1" ] ; then
 	[ ! -z "`pidof koolproxy`" ] && logger -t "【koolproxy】" "停止 koolproxy" && koolproxy_close
-	{ eval $(ps -w | grep "$scriptname" | grep -v grep | awk '{print "kill "$1";";}'); exit 0; }
+	{ kill_ps "$scriptname" exit0; exit 0; }
 fi
 if [ "$koolproxy_enable" = "1" ] ; then
 	if [ "$needed_restart" = "1" ] ; then
@@ -181,11 +216,11 @@ nvram set koolproxy_h="`/tmp/7620koolproxy/koolproxy -h | awk 'NR==1{print}'`】
 fi
 cat > "/tmp/sh_ad_kp_keey_k.sh" <<-ADMK
 #!/bin/sh
+source /etc/storage/script/init.sh
 sleep 919
 koolproxy_enable=\`nvram get koolproxy_enable\`
 if [ ! -f /tmp/cron_adb.lock ] && [ "\$koolproxy_enable" = "1" ] ; then
-eval \$(ps -w | grep "$scriptname" | grep -v grep | awk '{print "kill "\$1";";}')
-eval \$(ps -w | grep "$scriptname keep" | grep -v grep | awk '{print "kill "\$1";";}')
+kill_ps "$scriptname"
 eval "$scriptfilepath keep &"
 exit 0
 fi
@@ -197,9 +232,9 @@ killall -9 sh_ad_kp_keey_k.sh
 
 rm -f /tmp/cron_adb.lock
 reb="1"
-[ -z $ss_link_1 ] && ss_link_1="email.163.com" && nvram set ss_link_1="email.163.com"
+[ -z $ss_link_1 ] && ss_link_1="www.163.com" && nvram set ss_link_1="www.163.com"
 [ -z $ss_link_2 ] && ss_link_2="www.google.com.hk" && nvram set ss_link_2="www.google.com.hk"
-[ $ss_link_1 == "www.163.com" ] && ss_link_1="email.163.com" && nvram set ss_link_1="email.163.com"
+[ $ss_link_1 == "email.163.com" ] && ss_link_1="www.163.com" && nvram set ss_link_1="www.163.com"
 while true; do
 koolproxy_enable=`nvram get koolproxy_enable`
 [ "$koolproxy_enable" != "1" ] && exit
@@ -217,30 +252,32 @@ if [ ! -f /tmp/cron_adb.lock ] ; then
 		sleep 5
 		reboot
 	fi
-	hash check_network 2>/dev/null && {
-	check_network 3
-	[ "$?" == "0" ] && check=200 || { check=404;  sleep 3; }
+	check=0
+	hash check_network 2>/dev/null && check=1
+	if [ "$check" == "1" ] ; then
+		check_network 3
+		[ "$?" == "0" ] && check=200 || { check=404;  sleep 1; }
 		if [ "$check" == "404" ] ; then
 			check_network 3
 			[ "$?" == "0" ] && check=200 || check=404
 		fi
-	}
-	hash check_network 2>/dev/null || check=404
-	[ "$check" == "404" ] && {
-	curltest=`which curl`
-	if [ -z "$curltest" ] || [ ! -s "`which curl`" ] ; then
-		wget --continue --no-check-certificate -q -T 10 "$ss_link_1" -O /dev/null
-		[ "$?" == "0" ] && check=200 || { check=404;  sleep 3; }
-		if [ "$check" == "404" ] ; then
-			wget --continue --no-check-certificate -q -T 10 "$ss_link_1" -O /dev/null
-			[ "$?" == "0" ] && check=200 || check=404
-		fi
-	else
-		check=`curl -k -s -w "%{http_code}" "$ss_link_1" -o /dev/null`
-		[ "$check" != "200" ] &&  sleep 3
-		[ "$check" != "200" ] && check=`curl -k -s -w "%{http_code}" "$ss_link_1" -o /dev/null`
 	fi
-	}
+	hash check_network 2>/dev/null || check=404
+	if [ "$check" == "404" ] ; then
+		curltest=`which curl`
+		if [ -z "$curltest" ] || [ ! -s "`which curl`" ] ; then
+			wget --no-check-certificate -q -T 10 "$ss_link_1" -O /dev/null
+			[ "$?" == "0" ] && check=200 || { check=404;  sleep 1; }
+			if [ "$check" == "404" ] ; then
+				wget --no-check-certificate -q -T 10 "$ss_link_1" -O /dev/null
+				[ "$?" == "0" ] && check=200 || check=404
+			fi
+		else
+			check=`curl -k -s -w "%{http_code}" "$ss_link_1" -o /dev/null`
+			[ "$check" != "200" ] &&  sleep 1
+			[ "$check" != "200" ] && check=`curl -k -s -w "%{http_code}" "$ss_link_1" -o /dev/null`
+		fi
+	fi
 	if [ "$check" == "200" ] && [ ! -f /tmp/cron_adb.lock ] ; then
 		reb=1
 		PIDS=$(ps -w | grep "/tmp/7620koolproxy/koolproxy" | grep -v "grep" | wc -l)
@@ -309,7 +346,7 @@ done
 koolproxy_keepcpu () {
 if [ "$adbyby_CPUAverages" = "1" ] && [ ! -f /tmp/cron_adb.lock ] ; then
 	processor=`cat /proc/cpuinfo| grep "processor"| wc -l`
-	processor=`expr $processor \* 2`
+	[ "$processor" = "1" ] && processor=`expr $processor \* 2`
 	CPULoad=`uptime |sed -e 's/\ *//g' -e 's/.*://g' | awk -F ',' '{print $2;}' | sed -e 's/\..*//g'`
 	if [ $((CPULoad)) -ge "$processor" ] ; then
 		logger -t "【koolproxy】" "CPU 负载拥堵, 关闭 koolproxy"
@@ -318,8 +355,6 @@ if [ "$adbyby_CPUAverages" = "1" ] && [ ! -f /tmp/cron_adb.lock ] ; then
 		killall -15 koolproxy
 		killall -9 koolproxy
 		touch /tmp/cron_adb.lock
-		processor=`cat /proc/cpuinfo| grep "processor"| wc -l`
-		processor=`expr $processor \* 2`
 		while [[ "$CPULoad" -gt "$processor" ]] 
 		do
 			sleep 62
@@ -346,9 +381,9 @@ killall -15 koolproxy sh_ad_kp_keey_k.sh
 killall -9 koolproxy sh_ad_kp_keey_k.sh
 rm -f /tmp/adbyby_host.conf
 rm -f /tmp/7620koolproxy.tgz /tmp/cron_adb.lock /tmp/sh_ad_kp_keey_k.sh /tmp/cp_rules.lock
-eval $(ps -w | grep "_kool_proxy keep" | grep -v grep | awk '{print "kill "$1";";}')
-eval $(ps -w | grep "_kool_proxy.sh keep" | grep -v grep | awk '{print "kill "$1";";}')
-eval $(ps -w | grep "$scriptname keep" | grep -v grep | awk '{print "kill "$1";";}')
+kill_ps "/tmp/script/_kool_proxy"
+kill_ps "_kool_proxy.sh"
+kill_ps "$scriptname"
 }
 
 koolproxy_start () {
@@ -463,23 +498,23 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 			sleep 1
 	done
 	[ -s /tmp/7620koolproxy/data/rules/koolproxy.txt ] && nvram set koolproxy_uprules=2
-	[ ! -s /tmp/7620koolproxy/data/rules/koolproxy.txt ] && {
-	logger -t "【koolproxy】" "自动更新规则失效，启用脚本手动下载更新。"
-	nvram set koolproxy_uprules=1
-	mkdir -p /tmp/7620koolproxy/data/rules
-	cd /tmp/7620koolproxy/data/rules
-	hash daydayup 2>/dev/null && update_kp_rules_daydayup
-	hash daydayup 2>/dev/null || update_kp_rules
-	rules_nu="`cat /tmp/7620koolproxy/data/rules/koolproxy.txt | grep -v ! | wc -l`"
-	if [ $rules_nu -lt 5 ] ; then
-		logger -t "【koolproxy】" "错误！下载规则数 $rules_nu ，再次启用脚本手动下载更新。"
-		update_kp_rules 5
+	if [ ! -s /tmp/7620koolproxy/data/rules/koolproxy.txt ] ; then
+		logger -t "【koolproxy】" "自动更新规则失效，启用脚本手动下载更新。"
+		nvram set koolproxy_uprules=1
+		mkdir -p /tmp/7620koolproxy/data/rules
+		cd /tmp/7620koolproxy/data/rules
+		hash daydayup 2>/dev/null && update_kp_rules_daydayup
+		hash daydayup 2>/dev/null || update_kp_rules
+		rules_nu="`cat /tmp/7620koolproxy/data/rules/koolproxy.txt | grep -v ! | wc -l`"
+		if [ $rules_nu -lt 5 ] ; then
+			logger -t "【koolproxy】" "错误！下载规则数 $rules_nu ，再次启用脚本手动下载更新。"
+			update_kp_rules 5
+		fi
+		killall koolproxy
+		cd /tmp/7620koolproxy/
+		/tmp/7620koolproxy/koolproxy $mode_video -d # >/dev/null 2>&1 &
+		koolproxy_cron_job
 	fi
-	killall koolproxy
-	cd /tmp/7620koolproxy/
-	/tmp/7620koolproxy/koolproxy $mode_video -d # >/dev/null 2>&1 &
-	koolproxy_cron_job
-	}
 	hash krdl 2>/dev/null && krdl_ipset
 fi
 if [ -s /tmp/7620koolproxy/data/rules/koolproxy.txt ] ; then
@@ -595,6 +630,7 @@ eval $(ls| grep txt.http| awk '{print "cat /tmp/7620koolproxy/data/rules/"$1" >>
 # 提取IP
 cat  /tmp/7620koolproxy/domain.txt /tmp/7620koolproxy/koolproxy_blockip.txt | grep -Eo '^[0-9\.]*$' | sort -u | grep -v "^$" > /tmp/7620koolproxy/ip.txt
 cat /tmp/7620koolproxy/ip.txt /tmp/7620koolproxy/koolproxy_blockip.txt | sort -u > /tmp/7620koolproxy/koolproxy_blockip.txt
+sed -Ei '/0.0.0.0/d' /tmp/7620koolproxy/koolproxy_blockip.txt
 # 提取Domain
 cat  /tmp/7620koolproxy/domain.txt | grep  -Ev '^[0-9\.]*$' | sort -u > /tmp/7620koolproxy/domain2.txt
 sed -e "s/^/ipset=\/\./" -e "s/$/\/black_koolproxy/" -i /tmp/7620koolproxy/domain2.txt
@@ -975,7 +1011,7 @@ koolproxy_cron_job(){
 	[ $koolproxy_update_min -gt 59 ] && koolproxy_update_min=59 && nvram set koolproxy_update_min=$koolproxy_update_min
 	[ $koolproxy_update_min -lt 0 ] && koolproxy_update_min=0 && nvram set koolproxy_update_min=$koolproxy_update_min
 		logger -t "【koolproxy】" "开启规则定时更新，每隔"$koolproxy_update_inter_hour"时"$koolproxy_update_inter_min"分，检查在线规则更新..."
-		cru.sh a koolproxy_update "*/$koolproxy_update_min */$koolproxy_update_hour * * * $scriptfilepath update &" &
+		cru.sh a koolproxy_update "$koolproxy_update_min */$koolproxy_update_hour * * * $scriptfilepath update &" &
 	else
 		logger -t "【koolproxy】" "规则自动更新关闭状态，不启用自动更新..."
 	fi
@@ -996,14 +1032,14 @@ rm -f /tmp/arNslookup/$$
 else
 	curltest=`which curl`
 	if [ -z "$curltest" ] || [ ! -s "`which curl`" ] ; then
-		Address=`wget --continue --no-check-certificate --quiet --output-document=- http://119.29.29.29/d?dn=$1`
+		Address="`wget --no-check-certificate --quiet --output-document=- http://119.29.29.29/d?dn=$1`"
 		if [ $? -eq 0 ]; then
-		echo $Address |  sed s/\;/"\n"/g
+		echo "$Address" |  sed s/\;/"\n"/g | grep -E -o '([0-9]+\.){3}[0-9]+'
 		fi
 	else
-		Address=`curl -k http://119.29.29.29/d?dn=$1`
+		Address="`curl -k http://119.29.29.29/d?dn=$1`"
 		if [ $? -eq 0 ]; then
-		echo $Address |  sed s/\;/"\n"/g
+		echo "$Address" |  sed s/\;/"\n"/g | grep -E -o '([0-9]+\.){3}[0-9]+'
 		fi
 	fi
 fi
@@ -1012,11 +1048,82 @@ fi
 initopt () {
 optPath=`grep ' /opt ' /proc/mounts | grep tmpfs`
 [ ! -z "$optPath" ] && return
-if [ -z "$(echo $scriptfilepath | grep "/tmp/script/")" ] && [ -s "/opt/etc/init.d/rc.func" ] ; then
-	cp -Hf "$scriptfilepath" "/opt/etc/init.d/$scriptname"
+if [ ! -z "$(echo $scriptfilepath | grep -v "/opt/etc/init")" ] && [ -s "/opt/etc/init.d/rc.func" ] ; then
+	{ echo '#!/bin/sh' ; echo $scriptfilepath '"$@"' '&' ; } > /opt/etc/init.d/$scriptname && chmod 777  /opt/etc/init.d/$scriptname
 fi
 
 }
+
+initconfig () {
+
+koolproxy_rules_script="/etc/storage/koolproxy_rules_script.sh"
+if [ ! -f "$koolproxy_rules_script" ] || [ ! -s "$koolproxy_rules_script" ] ; then
+	cat > "$koolproxy_rules_script" <<-\EEE
+!  ******************************* koolproxy 自定义过滤语法简表 *******************************
+!  ------------------------ 规则基于adblock规则，并进行了语法部分的扩展 ------------------------
+!  ABP规则请参考https://adblockplus.org/zh_CN/filters，下面为大致摘要
+!  "!" 为行注释符，注释行以该符号起始作为一行注释语义，用于规则描述
+!  "@@" 为白名单符，白名单具有最高优先级，放行过滤的网站，例如:@@||taobao.com
+!  ------------------------------------------------------------------------------------------
+!  "*" 为字符通配符，能够匹配0长度或任意长度的字符串，该通配符不能与正则语法混用。
+!  "^" 为分隔符，可以是除了字母、数字或者 _ - . % 之外的任何字符。
+!  "~" 为排除标识符，通配符能过滤大多数广告，但同时存在误杀, 可以通过排除标识符修正误杀链接。
+!  注：通配符仅在 url 规则中支持，html 规则中不支持
+!  ------------------------------------------------------------------------------------------
+!  "|" 为管线符号，来表示地址的最前端或最末端
+!  "||" 为子域通配符，方便匹配主域名下的所有子域
+!  用法及例子如下：(以下等号表示等价于)
+!  ||xx.com          =  http://xx.com || http://*.xx.com
+!  ||http://xx.com   =  http://xx.com || http://*.xx.com
+!  ||https://xx.com  =  https://xx.com || https://*.xx.com
+!  |xx.com           =  http://xx.com
+!  |http://xx.com    =  http://xx.com
+!  |https://xx.com   =  https://xx.com
+!  xx.com            =  http://*xx.com
+!  http://xx.com     =  http://*xx.com
+!  https://xx.com    =  https://xx.com
+!  ------------------------------------------------------------------------------------------
+!  支持html规则语法，例如：
+!  ||fulldls.com##.tp_reccomend_banner
+!  ||torrentzap.com##.tp_reccomend_banner
+!  但不支持adblock规则中，逗号合并符写法，例如：
+!  fulldls.com,torrentzap.com##.tp_reccomend_banner
+!  应该写成推荐样式或以下样式：
+!  fulldls.com##.tp_reccomend_banner
+!  torrentzap.com##.tp_reccomend_banner
+!  ------------------------------------------------------------------------------------------
+!  文本替换语法：$s@匹配内容@替换内容@
+!  文本替换例子：|http://cdn.pcbeta.js.inimc.com/data/cache/common.js?$s@var $banner = @@
+!  重定向语法：$r@匹配内容@替换内容@
+!  重定向例子：|http://koolshare.cn$r@http://koolshare.cn/*@http://www.qq.com@
+!  注：文本替换语法及重定向语法中的匹配内容不仅支持通配符功能，而且额外支持以下功能
+!  支持通配符 * 和 ? ，? 表示单个字符
+!  支持全正则匹配，/正则内容/ 表示应用正则匹配
+!  正则替换：替换内容支持 $1 $2 这样的符号
+!  普通替换：替换内容支持 * 这样的符号，表示把命中的内容复制到替换的内容。（类似 $1 $2，但是 * 号会自动计算数字）
+!  ------------------------------------------------------------------------------------------
+!  koolporxy支持https过滤功能，但考虑到https过滤的效率问题，目前仅允许非常明确的过滤指令。
+!  未来将逐步开放模糊https的相关语法，与普通语法同步，敬请期待。
+!  ******************************************************************************************
+
+EEE
+	chmod 755 "$koolproxy_rules_script"
+fi
+
+koolproxy_rules_list="/etc/storage/koolproxy_rules_list.sh"
+if [ ! -f "$koolproxy_rules_list" ] || [ ! -s "$koolproxy_rules_list" ] ; then
+	cat > "$koolproxy_rules_list" <<-\EEE
+# 【添加为#注释行，会自动忽略】
+# 【adbyby规则翻译的koolproxy兼容规则 by <dsyo2008> http://koolshare.cn/thread-83553-1-1.html】
+#https://raw.githubusercontent.com/dsyo2008/lazy_for_koolproxy/master/lazy_kp.txt
+
+EEE
+	chmod 755 "$koolproxy_rules_list"
+fi
+
+}
+
+initconfig
 
 case $ACTION in
 start)
